add function strlcpy

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include <stdio.h>
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	size_t	i;
